Create README.md
FizzBuzz is a common algorithm asked during an interview in which programming is involved. The prompt typically asks the person to write a program that prints out the numbers 1-100 with every multiple of 3 being placed with "Fizz" and every multiple of 5 being replaced with "Buzz". If a number is a multiple of both 3 and 5, the output should be "FizzBuzz".